add: 멤버가 프로젝트에 참여하는 흐름 추가
BaseProjectWaitingActivity 로 중복 코드 관리
