+++ app/src/main/AndroidManifest.xml
@@ -28,13 +28,11 @@
             </intent-filter>
 
         </activity>
-        <activity android:name=".ui.AddProjectActivity" >
-
-
-        </activity>
+        <activity android:name=".ui.AddProjectActivity" />
         <activity
             android:name=".ui.MainActivity"
             android:windowSoftInputMode="adjustNothing"/>
+        <activity android:name=".ui.MemberProjectWaitingActivity"/>
         <activity android:name=".ui.HostRoundWaitingActivity"/>
         <activity android:name=".ui.ParticipatedProjectListActivity" />
         <activity android:name=".ui.ParticipatedProjectDetailActivity" />
