+++ app/src/main/java/com/stormers/storm/ui/MemberProjectWaitingActivity.kt
@@ -1,0 +1,12 @@
+package com.stormers.storm.ui
+
+import android.os.Bundle
+import android.os.PersistableBundle
+import com.stormers.storm.project.base.BaseProjectWaitingActivity
+
+class MemberProjectWaitingActivity : BaseProjectWaitingActivity(false) {
+
+    override fun onCreate(savedInstanceState: Bundle?, persistentState: PersistableBundle?) {
+        super.onCreate(savedInstanceState, persistentState)
+    }
+}
